fix: correctly chain CSS variable fallbacks
Unfortunately, the CSS spec does not allow all fallbacks to be chained together in a single var()
function. This means that if you want to use multiple fallbacks, you have to repeat the var()
function for each one. Definitely not ideal, but that's how it is.

See also: https://developer.mozilla.org/en-US/docs/Web/CSS/var#formal_syntax